Fix stack smash in sendBoard on 64-bit builds
The on-stack buf was sized using sizeof(int) per board cell, but each
cell is written via BTNET_PUTLONG, which copies sizeof(unsigned long)
bytes. On the original 32-bit Solaris these were both 4 bytes; on
64-bit Linux unsigned long is 8, so writing all 280 cells overruns
the buffer by ~1.1 KB and trips the stack canary. Triggered when
weapons like BT_AMES / BT_ACE / BT_CONDOR fire and the board goes
over the wire.

+++ usr/src/game/BTCommManager.C
@@ -349,7 +349,7 @@ void BTCommManager::sendWeapon(BTWeapon *weapon)
 void
 BTCommManager::sendBoard(BTBoard *board)
 {
-	char buf[(BT_BOARD_HGT * BT_BOARD_WTH * sizeof (int)) +
+	char buf[(BT_BOARD_HGT * BT_BOARD_WTH * sizeof (unsigned long)) +
 	    (3 * sizeof(unsigned short))];
 	char *bufptr = buf;
 	unsigned short ts;
